dsad

+++ GoldShop.Domain/Entity/Product/DiscountCode.cs
@@ -5,7 +5,7 @@ namespace GoldShop.Domain.Entity.Product;
 public class DiscountCode:BaseEntity
 {
     public string Code { get; set; } = string.Empty;
-    public double Amount { get; set; }
+    public long Amount { get; set; }
     public int Count { get; set; }
     public bool IsActive { get; set; }
 }
