+++ GoldShop.Domain/Entity/Product/GoldPrice.cs
@@ -5,8 +5,8 @@ namespace GoldShop.Domain.Entity.Product;
 
 public class GoldPrice : BaseEntity
 {
-    public double PricePerGram { get; set; }
-    public double PriceApi { get; set; }
+    public long PricePerGram { get; set; }
+    public long PriceApi { get; set; }
     public string GenderType { get; set; } = string.Empty; //نوع جنس
     public PriceType PriceType { get; set; }
     public DateTime UpdateTime { get; set; }
